make tolerant of partially complete batch jobs

+++ batch_combine/checkall.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 #
-# Run over all root files
+# Run over all specified root files 
+# testing to see whether there is a mismatch in the event counts 
+# that appears to be a symptom of a race condition.
 #
 
 echo $USER
@@ -13,40 +15,20 @@ rpath=/home/${USER}/slepton/ParallelSleptons/ExecutionDirectory
 for lfile in $(cat ${flistfile})
 
 do
-#   echo ' '
    name=$(echo "$lfile" | cut -f 1 -d '.')
-#   echo $name
-#   listfile=${name}.list
-
    rootfile=${rpath}/${name}/susyHists.root
    echo ${rootfile}
 
-#   root -l -b -q 'mismatch.C("\$rootfile")'
-
-#   root -l -b -q 'mismatch.C("/home/graham/slepton/ParallelSleptons/ExecutionDirectory/WJetsToLNu_HT-100To200_TuneCP5_13TeV-madgraphMLM-pythia8_Fall17_102X/susyHists.root")'
-
-#OK make this simple by making a local symbolic link
-
-   ln -s ${rootfile} MyFile.root
-
-   root -l -b -q 'mismatch.C("MyFile.root")'
-
-   rm MyFile.root
-
-#   cmd="/home/graham/slepton/ParallelSleptons/batch_combine/mismatch.C(\"${rootfile}\")"
-#   echo $cmd
-
-#   pwd
-
-#   cmd2="root -l -b -q '$cmd'"
-#   echo $cmd2
-
-#   $cmd2
-
-#   root -l -b -q "mismatch.C(\"All.root\")"
-
-    echo ' '
-    echo ' '
+#Make this simple by making a local symbolic link
+   if [ -s ${rootfile} ]
+   then
+ 
+       ln -s ${rootfile} MyFile.root
+       root -l -b -q 'mismatch.C("MyFile.root")'
+       rm MyFile.root
+   fi
+   echo ' '
+   echo ' '
 
 done
